src/requirements.txt
pandas
scikit-learn
python-binance
matplotlib
tabulate